Refactor ThemeSettings
