+++ site/src/App/ThemeSetting/ThemeSettingContext.tsx
@@ -8,7 +8,6 @@ import React, {
 import { useLocalStorage } from 'react-use';
 
 import * as themes from '../../../../lib/themes';
-import { BraidProvider, Box } from '../../../../lib/components';
 
 const defaultTheme = 'seekUnifiedBeta' as const;
 
@@ -31,21 +30,6 @@ export function useThemeSettings() {
   }
 
   return themeSettings;
-}
-
-interface ThemedExampleProps {
-  children: ReactNode;
-}
-export function ThemedExample({ children }: ThemedExampleProps) {
-  const { theme, ready } = useThemeSettings();
-
-  return (
-    <Box style={{ opacity: ready ? 1 : 0 }} transition="fast">
-      <BraidProvider styleBody={false} theme={themes[theme]}>
-        {children}
-      </BraidProvider>
-    </Box>
-  );
 }
 
 interface ThemeSettingProviderProps {
