+++ site/src/App/ThemeSetting/index.ts
@@ -1,0 +1,7 @@
+export { ThemedExample } from './ThemedExample';
+export {
+  ThemeSettingProvider,
+  ThemeKey,
+  useThemeSettings,
+} from './ThemeSettingContext';
+export { ThemeToggle } from './ThemeToggle';
